Show first five search inputs
"A good use for array methods like push() and pop() is when you are maintaining a record of currently active items in a web app. As new objects are created and added to the array, older ones can be deleted from the array to maintain the desired number.

In this example we're going to show a much simpler use — here we're giving you a fake search site, with a search box. The idea is that when terms are entered in the search box, the top 5 previous search terms are displayed in the list. When the number of terms goes over 5, the last term starts being deleted each time a new term is added to the top, so the 5 previous terms are always displayed."

https://developer.mozilla.org/en-US/docs/Learn/JavaScript/First_steps/Arrays