16 different color blocks to click on
"e.target is incredibly useful when you want to set the same event handler on multiple elements, and do something to all of them when an event occurs on them. You might for example have a set of 16 tiles that disappear when they are clicked on. It is useful to always be able to just set the thing to disappear as e.target, rather than having to select it in some more difficult way. In [this example] we create 16 <div> elements using JavaScript. We then select all of them using document.querySelectorAll(), then loop through each one, adding an onclick handler to each that makes it so that a random color is applied to each one when clicked."
https://developer.mozilla.org/en-US/docs/Learn/JavaScript/Building_blocks/Events